[gui-tests] Separate steps (#10381)
* let squish search the test cases

* remove unnecessary config files

* separate server context

* separate sharing context

* separate account context

* merge socket helper to sync helper

* separate sync context

* refactor step to file_context
