--- test/gui/config.xml
@@ -1,7 +1,0 @@
-<testconfig version="1.0">
-<information>
-<summary/>
-<description/>
-</information>
-<testsettings/>
-</testconfig>
