+++ test/gui/suite.conf
@@ -1,9 +1,7 @@
 AUT=owncloud --showsettings
 ENVVARS=envvars
 HOOK_SUB_PROCESSES=false
-IMPLICITAUTSTART=0
+IMPLICITAUTSTART=false
 LANGUAGE=Python
 OBJECTMAPSTYLE=script
-TEST_CASES=tst_addAccount tst_sharing tst_syncing tst_loginLogout tst_removeAccountConnection tst_checkAlltabs tst_vfs tst_deletFilesFolders tst_editFiles
-VERSION=3
 WRAPPERS=Qt
